+++ backend/src/controllers/endpointsController.js
@@ -7,8 +7,8 @@ const dbConfig = {
     host:process.env.DB_HOST,
     port:process.env.DB_PORT,
     user:process.env.DB_USER,
-    pass:process.env.DB_PASS,
-    name:process.env.DB_NAME
+    password:process.env.DB_PASSWORD,
+    database:process.env.DB_NAME
 };
 
 let db;