+++ backend/src/models/DbObject.js
@@ -1,61 +1,93 @@
-// import required files
 const mysql = require("mysql2/promise");
 
 /**
- * Class representing a Database Object.
+ * Represents a database object that manages the connection, querying, and disconnection
+ * to a MySQL database using the mysql2/promise library.
+ *
+ * @class DbObject
+ * @constructor
+ * @param {Object} config - The configuration object for the MySQL connection.
  */
 class DbObject{
-    constructor(config) {           // Constructor for DbObject class, takes a configuration object as a parameter
-        this.config = config;       // Store the provided configuration object in the 'config' property
-        this.connection = null;     // Initialize the 'connection' property to null
+    /**
+     * Creates an instance of DbObject.
+     *
+     * @constructor
+     * @param {Object} config - The configuration object for the MySQL connection.
+     */
+    constructor(config){
+        /**
+         * The configuration object for the MySQL connection.
+         *
+         * @property {Object} config
+         */
+        this.config = config;
+        /**
+         * The MySQL database connection object.
+         *
+         * @property {Object} connection
+         */
+        this.connection = null;
     }
 
     /**
-     * Establishes a connection to the database.
-     * @throws {Error} Throws an error if the connection cannot be established.
+     * Establishes a connection to the MySQL database using the provided configuration.
+     *
+     * @async
+     * @method connect
+     * @throws {Error} Throws an error if the connection attempt fails.
      */
-    async connect() {
-        try {
-        this.connection = await mysql.createConnection(this.config);
-        } catch(error) {
+    async connect(){
+        try{
+            this.connection = await mysql.createConnection(this.config);
+        }
+        catch(error){
             throw error;
         }
     }
 
     /**
-     * Executes a prepared query on the database.
-     * @param {string} query - SQL query to be executed.
-     * @param {Array} bindings - Array of values to bind to the query.
-     * @returns {Promise<Array>} - A Promise that resolves to the result of the query.
+     * Executes a SQL query on the connected database.
+     *
+     * @async
+     * @method query
+     * @param {string} query - The SQL query to execute.
+     * @param {Array} bindings - An array of parameter values to bind to the query.
      * @throws {Error} Throws an error if the query execution fails.
      */
-    async query(query, bindings) {
+    async query(query, bindings){
         try {
-            if(!this.connection || this.connection.state === "disconnected") {
-            await this.connect();
+            if (!this.connection || this.connection.state === 'disconnected') {
+                await this.connect();
             }
+
             return await this.connection.execute(query, bindings);
-        } catch(error) {
+        }
+        catch(error){
             throw error;
         }
     }
 
     /**
-     * Ends the database connection if it is active.
-     * @throws {Error} Throws an error if there is a problem ending the connection.
+     * Ends the current database connection.
+     *
+     * @async
+     * @method end
+     * @throws {Error} Throws an error if there is an issue ending the connection.
      */
-    async end() {
-        try {
+    async end(){
+        try{
             if (this.connection && this.connection.state !== 'disconnected') {
                 await this.connection.end();
             }
-        } catch(error) {
+        }
+        catch(error){
             throw error;
-        } finally{
+        }
+        finally{
             this.connection = null;
         }
     }
-
 }
 
 module.exports = DbObject;