+++ frontend/src/components/NavBar/NavBar.tsx
@@ -1,51 +1,19 @@
-import { FaShoppingBag } from "react-icons/fa";
-import { NavLink } from "react-router-dom";
-import { useCartContext } from "../../contexts/CartContext";
-/**
- * NavBar component representing the navigation bar.
- * @component
- * @returns {JSX.Element} - NavBar component
- */
-const NavBar: React.FC = () => {
+import { Link } from "react-router-dom"
+import { useCartContext } from "../../contexts/CartContext"
 
+const Navbar: React.FC = () => {
     const {
-        handleToggleCartModalClick
-      } = useCartContext();
-
-    return (
-        <nav id="main-navbar">
-            <li className="li-nav">
-                <NavLink
-                    to="/"
-                    className={({ isActive, isPending }) =>
-                        isPending ? "pending" : isActive ? "active" : ""
-                    }
-                    style={{
-                        textDecoration: 'none',
-                        color: 'var(--main-text-color)',
-                    }}
-                    >
-                    Hem
-                </NavLink>
-            </li>
-            <li className="li-nav">
-                <NavLink
-                    to="/om-oss"
-                    className={({ isActive, isPending }) =>
-                        isPending ? "pending" : isActive ? "active" : ""
-                    }
-                    style={{
-                        textDecoration: 'none',
-                        color: 'var(--main-text-color)',
-                    }}
-                    >
-                    Om oss
-                </NavLink>
-            </li>
-            <FaShoppingBag id="cart-icon" onClick={handleToggleCartModalClick}/>
-            {/* onClick={handleOpenCartModalClick} */}
-        </nav>
-    );
+        setIsCartModalOpen
+    } = useCartContext()
+    return(
+        <>
+        <label className="navButton"><Link to="/">Shop</Link></label>
+        <label className="navButton"><Link to="/cart">Cart</Link></label>
+        <label className="navButton"><Link to="/checkout" onClick={() => setIsCartModalOpen(true)}>Checkout</Link></label>
+        <label className="navButton"><Link to="/orders">Orders</Link></label>
+        <label className="navButton"><Link to="/about/policies">About Us</Link></label>
+        </>
+    )
 }
 
-export default NavBar
+export default Navbar