+++ frontend/src/pages/AboutUs/AboutUs.tsx
@@ -1,0 +1,9 @@
+import DisclaimerPanel from "../../components/DisclaimerPanel/DisclaimerPanel"
+
+const AboutUs = () => {
+    return (
+        <DisclaimerPanel />
+    );
+}
+
+export default AboutUs