+++ frontend/src/pages/Orders/Orders.tsx
@@ -1,7 +1,13 @@
-import { OrderContextProvider } from "../../contexts/OrderContext";
-import OrderModal from "../../components/OrderModal/OrderModal";
+import { PropsWithChildren } from "react"
+import { OrderContextProvider } from "../../contexts/OrderContext"
+import OrderModal from "../../components/OrderModal/OrderModal"
 
-const Orders: React.FC = () => {
+/**
+ * Orders component responsible for rendering the order modal.
+ * This component wraps the OrderModal component with OrderContextProvider.
+ * @returns {JSX.Element} JSX for the Orders component.
+ */
+const Orders: React.FC<PropsWithChildren<{}>> = () => {
     return (
         <OrderContextProvider>
             <OrderModal />