+++ docs/changelog.md
@@ -9,5 +9,11 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2024-02-02
 
 ### Added
-- Initial release
+- Initial release of the webshop application.
+
+  Features:
+  - Product listing and details.
+  - Filtering and sorting functionality.
+  - Shopping cart functionality.
+  - Checkout process.